Initial Cleanup of files
