--- gurobi.log
@@ -1,6 +1,0 @@
-
-Gurobi 9.0.0 (linux64, Python) logging started Mon Dec  2 22:33:55 2019
-
-Using license file /home/bhvrddy_gmail_com/gurobi.lic
-Set parameter LogFile to value gurobi.log
-Academic license - for non-commercial use only
