--- outputs/200/200_November24_1059.out
@@ -1,67 +1,0 @@
-0 1 6 11 16 21 26 32 33 31 37 36 42 41 46 51 57 58 56 61 66 71 76 81 87 86 92 93 91 96 101 106 112 111 116 121 127 128 126 131 136 141 146 151 156 162 163 161 167 166 171 176 181 186 192 191 196 197 198 0 164 161 129 130 126 105 104 101 102 103 101 95 94 91 59 60 56 52 53 51 35 31 30 29 26 25 24 21 12 13 11
-65
-1 3
-6 7
-11 14
-12 12
-13 13
-16 17 19
-21 22
-24 24
-25 25
-26 27
-29 29
-30 30
-32 32
-33 33
-35 35 34
-36 39
-37 38
-41 44
-42 42
-46 48 47
-52 52
-53 53
-58 58
-59 59
-60 60
-61 63 65 64 62
-66 70 68
-71 73 72 75 74
-76 78 79 80
-81 83 84 82
-86 90
-87 87
-93 93
-94 94
-95 95
-96 100 99 98 97
-102 102
-103 103
-104 104
-105 105
-106 109 107 108
-111 114 113
-112 112
-116 118
-121 122
-127 127
-128 128
-129 129
-130 130
-131 135
-136 137 139 140 138
-141 144 143 145 142
-146 149
-151 155 152 153
-156 159
-162 162
-163 163
-164 164 165
-166 169 170
-171 173
-176 180 179
-181 183 184 185
-186 189 187
-191 193 195
-192 192
