+++ tests/inputs/tests/modified_hourglass.in
