+++ tests/inputs/tests/triple_hourglass.in
