+++ tests/inputs/tests/walk_home_hourglass.in
