+++ tests/outputs/6/multiple.out
